fixing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5908 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSInternalJunction.cpp
@@ -74,7 +74,7 @@ void
 MSInternalJunction::postloadInit()
 {
     // inform links where they have to report approaching vehicles to
-    size_t requestPos = 0;
+    unsigned int requestPos = 0;
     if (myIncomingLanes.size()!=0) {
         // for the first incoming lane
         const MSLinkCont &links = myIncomingLanes[0]->getLinkCont();
